[refactor] move the twisted http code to leap.common
